+++ ui/packages/shared/profile/src/ProfileExplorer/ProfileExplorerSingle.tsx
@@ -11,8 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import {NavigateFunction} from '@parca/functions';
 import {QueryServiceClient} from '@parca/client';
-import {ProfileSelection, ProfileViewWithData, NavigateFunction} from '..';
+import {ProfileSelection, ProfileViewWithData} from '..';
 
 import ProfileSelector, {QuerySelection} from '../ProfileSelector';
 interface ProfileExplorerSingleProps {