+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.33](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.32...@parca/grafana-panel@0.0.33) (2022-12-12)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.32](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.31...@parca/grafana-panel@0.0.32) (2022-12-12)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.31 (2022-12-10)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.30 (2022-12-08)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.29](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.28...@parca/grafana-panel@0.0.29) (2022-12-07)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.28](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.27...@parca/grafana-panel@0.0.28) (2022-12-06)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## [0.0.27](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.26...@parca/grafana-panel@0.0.27) (2022-11-29)
 
 **Note:** Version bump only for package @parca/grafana-panel